+++ Tools/xmlTools.py
@@ -27,14 +27,13 @@ def ReadXml(i_filePath):
             defaultConfigFile = os.path.join(gs._FOLDER_TEMPLATES, gs._FILE_CONFIG+'.default')
             with open(defaultConfigFile, 'r') as src, open(gs._FILE_CONFIG, 'w') as dst:
                 dst.write(src.read())
-        elif i_filePath == os.path.join(gs.configList['codes-folder'], gs._FILE_CONTROL):            # CONTROL FILE corruption
+        elif i_filePath == gs._FILE_CONTROL:            # CONTROL FILE corruption
             error = f'Your {gs._FILE_CONTROL} is corrupted, we are going to attempt to regenerate it'
-            fileControl = os.path.join(gs.configList['codes-folder'], gs._FILE_CONTROL)
-            if os.path.exists(fileControl):                  # Delete the corrupted control file and create a new one
-                os.remove(fileControl)
-            with open(fileControl, 'w') as file:
+            if os.path.exists(gs._FILE_CONTROL):                  # Delete the corrupted control file and create a new one
+                os.remove(gs._FILE_CONTROL)
+            with open(gs._FILE_CONTROL, 'w') as file:
                 file.write('<?xml version="1.0" ?><Root></Root>')
-        elif i_filePath[len(gs.configList['codes-folder'])+1:len(gs.configList['codes-folder'])+11] == 'output_log':    # CODE FILE corruption
+        elif i_filePath[len(gs._FOLDER_CODES)+1:len(gs._FOLDER_CODES)+11] == 'output_log':    # CODE FILE corruption
             fileCode = i_filePath
             error = f'{fileCode} is corrupted, we are going to attempt to regenerate it'
             if os.path.exists(fileCode):                    # Delete the corrupted code file
@@ -42,7 +41,7 @@ def ReadXml(i_filePath):
             allLogs = GetAllFiles(gs.configList['vrchat-log-path']+'output_log_*.txt')                  # I want control to forget the corrupted file
             fileCodeXML = os.path.join(gs.configList['vrchat-log-path'] ,fileCode.replace('Codes/','').replace('.xml','.txt'))
             allLogs.remove(fileCodeXML)
-            CleanControlEntries(os.path.join(gs.configList['codes-folder'], gs._FILE_CONTROL), allLogs) # And we do that here
+            CleanControlEntries(gs._FILE_CONTROL, allLogs) # And we do that here
             with open(fileCode, 'w') as file:
                 file.write('<?xml version="1.0" ?><Root></Root>')
             gs.forceRefreshCodes = True                     # We need to tell CodesHunter to refresh and read the Codes again